+++ changelog.md
@@ -9,6 +9,8 @@
 
 ## Bug fixes
 
+* fixed a bug in `from_neo4j`, where by default the node size would always be set if a `size` property was set on the node.
+
 
 ## Improvements